tools/assamilator/requirements.txt
pyudev
unsilence
python-dotenv
